Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
pymysql
python-dotenv


